Reorganise folders
Create a new folder called display and put the display functions inside it.
Move the implementTunnel file into the boundaries folder.
